# Atividade Final/docs/yaml/championships-api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
...
  /api/v1/championships/{championshipId}/seasons/{seasonId}/matches:
    get:
      tags:
      - match-controller
      summary: Get all registered matches
      operationId: getAll_3
      parameters:
      - name: championshipId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: seasonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "404":
          description: No matches found
        "200":
          description: List of all registered matches
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResponse'
    post:
      tags:
      - match-controller
      summary: Register a new match
      operationId: register_2
      parameters:
      - name: championshipId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: seasonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchRegisteringRequest'
        required: true
      responses:
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        "409":
          description: "Match is already registered \t\n The provided championship\
            \ does not exist \t\n The provided season does not exist"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictResponse'
        "201":
          description: Match registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResponse'
...
  /api/v1/divisions/{divisionId}:
    get:
      tags:
      - division-controller
      summary: Get a division by its id
      operationId: getOne
      parameters:
      - name: divisionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Found the division
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DivisionResponse'
        "404":
          description: Division not found
    delete:
      tags:
      - division-controller
      summary: Delete a division by its id
      operationId: deleteOne
      parameters:
      - name: divisionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Division deleted successfully
    patch:
      tags:
      - division-controller
      summary: Update a division by its id
      operationId: updateOne
      parameters:
      - name: divisionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DivisionUpdatingRequest'
        required: true
      responses:
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        "200":
          description: Division updated successfully
        "409":
          description: "The provided name is already in use \t\n The provided division\
            \ id does not exist"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictResponse'
  /api/v1/championships/{championshipId}:
    get:
      tags:
      - championship-controller
      summary: Get a championship by its id
      operationId: getOne_1
      parameters:
      - name: championshipId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Found the championship
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChampionshipResponse'
        "404":
          description: Championship not found
    delete:
      tags:
      - championship-controller
      summary: Delete a championship by its id
      operationId: deleteOne_1
      parameters:
      - name: championshipId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Championship deleted successfully
    patch:
      tags:
      - championship-controller
      summary: Update a championship by its id
      operationId: updateOne_1
      parameters:
      - name: championshipId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChampionshipUpdatingRequest'
        required: true
      responses:
        "400":
          description: Invalid payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestResponse'
        "200":
          description: Championship updated successfully
        "409":
          description: "The provided name is already in use \t\n The provided championship\
            \ id does not exist"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictResponse'
  /api/v1/championships/{championshipId}/seasons/{seasonId}/matches/{matchId}:
    get:
      tags:
      - match-controller
      summary: Get a match by its id
      operationId: getOne_3
      parameters:
      - name: championshipId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: seasonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "404":
          description: Match not found
        "200":
          description: Found the match
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MatchResponse'
    delete:
      tags:
      - match-controller
      summary: Delete a match by its id
      operationId: deleteOne_3
      parameters:
      - name: championshipId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: seasonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Match deleted successfully
    patch:
      tags:
      - match-controller
      summary: Update a match by its id
      operationId: updateOne_2
      parameters:
      - name: championshipId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: seasonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MatchUpdatingRequest'
        required: true
      responses:
        "409":
          description: The provided match id does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConflictResponse'
        "200":
          description: Match updated successfully
...
  /api/v1/championships/{championshipId}/seasons/{seasonId}:
    get:
      tags:
      - season-controller
      summary: Get a season by its id and championship id
      operationId: getOne_2
      parameters:
      - name: championshipId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: seasonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: Found the season
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SeasonResponse'
        "404":
          description: Season not found
    delete:
      tags:
      - season-controller
      summary: Delete a season by its id
      operationId: deleteOne_2
      parameters:
      - name: championshipId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: seasonId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: Season deleted successfully
...
components:
  schemas:
...
    ConflictResponse:
      type: object
      properties:
        message:
          type: string
        resource:
          type: object
    BadRequestResponse:
      type: object
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationResult'
    ValidationResult:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    DivisionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
...
    ChampionshipResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        division:
          $ref: '#/components/schemas/DivisionResponse'
        country:
          type: string
        seasons:
          $ref: '#/components/schemas/CollectionModelEntityModelSeasonResponse'
    CollectionModelEntityModelSeasonResponse:
      type: object
      properties:
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
        content:
          type: array
          items:
            $ref: '#/components/schemas/EntityModelSeasonResponse'
    EntityModelSeasonResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        year:
          type: integer
          format: int32
        matches:
          type: array
          items:
            $ref: '#/components/schemas/MatchResponse'
        started-at:
          type: string
          format: date-time
        ended-at:
          type: string
          format: date-time
        links:
          type: array
          items:
            $ref: '#/components/schemas/Link'
    Link:
      type: object
      properties:
        rel:
          type: string
        href:
          type: string
        hreflang:
          type: string
        media:
          type: string
        title:
          type: string
        type:
          type: string
        deprecation:
          type: string
        profile:
          type: string
        name:
          type: string
    MatchResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        arena:
          type: string
        date:
          type: string
          format: date-time
        team-id-a:
          type: string
          format: uuid
        team-id-b:
          type: string
          format: uuid
        created-at:
          type: string
          format: date-time
    SeasonResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        year:
          type: integer
          format: int32
        matches:
          type: array
          items:
            $ref: '#/components/schemas/MatchResponse'
        started-at:
          type: string
          format: date-time
        ended-at:
          type: string
          format: date-time
    MatchRegisteringRequest:
      required:
      - arena
      - date
      - team-id-a
      - team-id-b
      type: object
      properties:
        arena:
          type: string
        date:
          type: string
          format: date-time
        team-id-a:
          type: string
          format: uuid
        team-id-b:
          type: string
          format: uuid
...
    DivisionUpdatingRequest:
      required:
      - new-name
      type: object
      properties:
        new-name:
          type: string
    ChampionshipUpdatingRequest:
      type: object
      properties:
        new-name:
          type: string
    MatchUpdatingRequest:
      type: object
      properties:
        arena:
          type: string
        date:
          type: string
          format: date-time
        team-id-a:
          type: string
          format: uuid
        team-id-b:
          type: string
          format: uuid